% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amniota}
\alias{amniota}
\title{Amniote life-history traits}
\format{An object of class \code{data.frame} with 21322 rows and 36 columns.}
\usage{
amniota
}
\description{
An amniote life-history database to perform comparative analyses with birds, mammals, and reptiles
}
\seealso{
Other rawdata: \code{\link{amphibio}},
  \code{\link{arthropodtraits}}, \code{\link{carabids}},
  \code{\link{heteropteraRaw}}, \code{\link{mammaldiet}},
  \code{\link{pantheria}}, \code{\link{passerines}}
}
\keyword{datasets}
