% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amphibio}
\alias{amphibio}
\title{AmphiBIO, a global database for amphibian ecological traits}
\format{An object of class \code{data.frame} with 6776 rows and 38 columns.}
\source{
Cite as:
\itemize{
\item Oliveira, B. F. et al. AmphiBIO, a global database for amphibian ecological traits. Sci. Data. 4:170123 doi: 10.1038/sdata.2017.123 (2017). \url{https://doi.org/10.6084/m9.figshare.4644424};
}

\href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons BY 4.0}. You must give appropriate credit, provide a link to the license, and indicate if changes were made. You may do so in any reasonable manner, but not in any way that suggests the licensor endorses you or your use.
}
\usage{
amphibio
}
\description{
AmphiBIO, a global database for amphibian ecological traits
}
\seealso{
Other rawdata: \code{\link{amniota}},
  \code{\link{arthropodtraits}}, \code{\link{carabids}},
  \code{\link{heteroptera_raw}}, \code{\link{mammaldiet}},
  \code{\link{pantheria}}
}
\author{
Brunno Freire Oliveira, Vinícius Avelar São-Pedro, Georgina Santos-Barrera, Caterina Penone, and Gabriel C. Costa
}
\keyword{datasets}
