% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesaurus.R
\name{as.thesaurus}
\alias{as.thesaurus}
\title{Create a lookup table of trait definitions (a thesaurus of traits)}
\usage{
as.thesaurus(...)
}
\arguments{
\item{...}{}

\item{x}{}

\item{replace}{}

\item{appliestoall}{}
}
\value{
a list of formalized objects of class 'trait', as returned by function as.trait().
}
\description{
Create a lookup table of trait definitions (a thesaurus of traits)
}
\details{
the object class 'trait' comprises necessary information to map a trait name to a  trait definition, a target unit and a globally unique identifier. The thesaurus will be used in function 'standardize.traits()' to apply unit conversion and factor level harmonization.
}
\examples{

as.thesaurus(body_length = as.trait("body_length", traitUnit = "mm", traitUnitStd = "mm"), 
             antenna_length = as.trait("antenna_length", traitUnitStd = "mm"),
             metafemur_length = as.trait("metafemur_length"),
             eyewidth = as.trait("eyewidth_corr")
)

as.thesaurus(data.frame(
  traitName = c("body_length", "antenna_length", "metafemur_length", "eyewidth_corr"),
  # traitNameStd = traitName,
  traitUnit = "mm",
  traitType = "numeric",
  traitDescription = c("body length in mm", "length of antenna in mm", "length of metafemur in mm", "eye width in mm")
)
)
}
