% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesaurus.R
\name{as.trait}
\alias{as.trait}
\title{Create a trait definition}
\usage{
as.trait(traitName, traitID = NA, author = NA, URI = NA, traitUnit = NA,
  traitUnitStd = traitUnit, factorLevels = NA, traitType = NA,
  traitDescription = NA, bibliographicReference = NA, version = NA, ...)
}
\arguments{
\item{traitName}{}

\item{traitID}{}

\item{author}{}

\item{URI}{}

\item{traitUnit}{}

\item{factorLevels}{}

\item{traitDescription}{}

\item{bibliographicReference}{}

\item{version}{}

\item{...}{}
}
\description{
Create a trait definition
}
\examples{

body_size = as.trait("body_size", traitID = "http://url.to/thesaurus.html#body_size", author = "F.D. Schneider", traitDescription = "body length from head  to end of abdomen in mm")

}
