% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_factor_clocale}
\alias{as_factor_clocale}
\title{Locale-independent factor conversion}
\usage{
as_factor_clocale(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the same vector, but with any factorial content or colums with repeating character strings converted to factors, applying a locale-independent sorting.
}
\description{
Locale-independent factor conversion
}
\note{
Only for internal application in \code{\link[=as.traitdata]{as.traitdata()}} This deals with the problem described by Kurt Hornik \href{https://developer.r-project.org/Blog/public/2020/02/16/stringsasfactors/index.html}{(here)} for changes coming in R 4.0.0. In traitdataform, this concerns the locale-dependence of taxon and trait name sorting when calling published trait-datasets. For traits, the sorting order of factor levels will be superimposed by the order given in the thesaurus specification (if provided).
}
\keyword{internal}
