% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif_taxonomy.R
\name{get_gbif_taxonomy}
\alias{get_gbif_taxonomy}
\title{Get accepted canonical names and taxonomy for a given species name}
\usage{
get_gbif_taxonomy(x, infraspecies = FALSE, fuzzy = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{a character string or vector of species names.}

\item{infraspecies}{logical. If TRUE, the given name is resolved to infraspecies taxon, i.e. subspecies names will not be pooled (not working!)}

\item{fuzzy}{logical. If TRUE, function tries fuzzy matching for species requests. May produce output, if normal matching fails. (not working!)}

\item{verbose}{logical. If FALSE, warnings and messages are suppressed.}
}
\value{
a data.frame mapping the user supplied names to the accepted taxon names and higher taxonomic information (kingdom, phylum, class, order, family, genus). If 'synonym' is returned as TRUE, the user provided name has been identified as a synonym and was mapped to an accepted name. The field confidence reports the confidence of the matching procedure performed by the function 'get_gbifid_()' of the package 'taxize'. The taxonID is a globally valid URI that links to the taxon description of the GBIF backbone taxonomy.
}
\description{
Get accepted canonical names and taxonomy for a given species name
}
\examples{

get_gbif_taxonomy(c("Chorthippus albomarginatus", "Chorthippus apricarius", "Chorthippus biguttulus", "Chorthippus dorsatus", "Chorthippus montanus", "Chorthippus parallelus", "Chrysochraon dispar", "Conocephalus dorsalis", "Conocephalus fuscus", "Decticus verrucivorus", "Euthystira brachyptera", "Gomphocerippus rufus", "Gryllus campestris", "Metrioptera roeselii", "Omocestus viridulus", "Phaneroptera falcata", "Platycleis albopunctata", "Spec", "Stenobothrus lineatus", "Stenobothrus stigmaticus", "Stethophyma grossum", "Tetrix kraussi", "Tetrix subulata", "Tetrix tenuicornis", "Tetrix undulata", "Tettigonia cantans", "Tettigonia viridissima"))

}
