% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\docType{data}
\name{glossary}
\alias{glossary}
\title{Ecological Trait-data Standard vocabulary (ETS)}
\format{
An object of class \code{data.frame} with 80 rows and 13 columns.
}
\usage{
glossary
}
\description{
The terms and concepts as defined by the ETS (https://terminologies.gfbio.org/terms/ets/pages/)
}
\keyword{datasets}
