% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heteroptera_raw}
\alias{heteroptera_raw}
\alias{heteroptera}
\title{Heteroptera morphometry traits}
\format{An object of class \code{NULL} of length 0.}
\source{
\url{https://figshare.com/articles/Data_Paper_Data_Paper/3561936};
\href{http://creativecommons.org/publicdomain/zero/1.0/}{Creative Commons
0}. To the extent possible under law, the authors have waived all copyright
and related or neighboring rights to this data.
}
\usage{
heteroptera_raw

heteroptera
}
\value{
The dataset \code{heteropteraRaw} contains multiple observations of
each species (occurence table). The dataset \code{heteroptera} is a
compiled species-trait matrix.
}
\description{
The data sets comprise morphometric trait data of species that
were sampled and measured within the a subproject of the Biodiversity
Exploratories Project (Fischer et al. 2010) which focuses on the  effect of
land use on arthropod community composition and related processes in three
regions of Germany (Gossner et al. 2014, Simons et al. 2014, Simons et al.
2015).
}
\details{
Morphometric measures of Heteroptera sampled in grasslands across three
regions of Germany.

Cite this dataset as: \itemize{ \item M. Gossner, Martin; K. Simons,
Nadja; Höck, Leonhard; W. Weisser, Wolfgang (2016): Morphometric measures
of Heteroptera sampled in grasslands across three regions of Germany.
figshare. https://doi.org/10.6084/m9.figshare.c.3307611.v1 }
}
\seealso{
Other rawdata: \code{\link{amniota}},
  \code{\link{amphibio}}, \code{\link{arthropodtraits}},
  \code{\link{carabids}}, \code{\link{mammaldiet}},
  \code{\link{pantheria}}, \code{\link{pulldata}}

Other rawdata: \code{\link{amniota}},
  \code{\link{amphibio}}, \code{\link{arthropodtraits}},
  \code{\link{carabids}}, \code{\link{mammaldiet}},
  \code{\link{pantheria}}, \code{\link{pulldata}}
}
\author{
Martin M. Gossner , Nadja K. Simons, Leonhard Höck, Wolfgang W.
Weisser
}
\concept{rawdata}
\keyword{datasets}
