% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.traitdata.R
\name{mutate.traitdata}
\alias{mutate.traitdata}
\title{Mutate traits within a traitdata object.}
\usage{
mutate.traitdata(.data, ...)
}
\arguments{
\item{.data}{the traitdata object to transform}

\item{...}{named parameters giving definitions of new columns.}
}
\value{
an updated traitdata object with the new trait measures or facts
  appended to the original table. If the given trait name has been refined,
  it will be replaced.
}
\description{
This function allows to transform, factorize, or combine trait measurements
into compound measurements or update factor levels into binaries.
}
\details{
Note that units will not be maintained or refined. They can be
  assigned later through the \code{thesaurus()} function.
}
\examples{

dataset3 <- as.traitdata(arthropodtraits,
                         taxa = "SpeciesID",
                         traits = c("Body_Size", "Dispersal_ability", "Feeding_guild","Feeding_guild_short", "Feeding_mode", "Feeding_specialization", "Feeding_tissue", "Feeding_plant_part", "Endophagous_lifestyle", "Stratum_use", "Stratum_use_short"), 
                         units = c(Body_Size = "mm", Dispersal_ability = "unitless"),
                         keep = c(measurementRemark = "Remark")
)
head(dataset3)

updated <- mutate.traitdata(dataset3, predator = Feeding_guild == "c" )
head(updated)
levels(updated$traitName)
}
