% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pantheria}
\alias{pantheria}
\title{PanTHERIA mammal traits}
\format{An object of class \code{data.frame} with 5416 rows and 55 columns.}
\source{
\url {http://esapubs.org/archive/ecol/E090/184/metadata.htm}; \href{http://creativecommons.org/publicdomain/zero/1.0/}{Creative Commons 0}. To the extent possible under law, the authors have waived all copyright and related or neighboring rights to this data.
}
\usage{
pantheria
}
\description{
PanTHERIA: a species-level database of life history, ecology, and geography of extant and recently extinct mammals.
}
\details{
Data were gathered from the literature for 25 types of ecological and life history information for any extant or recently extinct species within class Mammalia (100740 data lines): 1. Activity Cycle; 2. Age at Eye Opening; 3. Age at First Birth; 4. Average Lifespan; 5. Body Mass; 6. Diet; 7. Dispersal Age; 8. Adult Limb Length; 9. Gestation Length; 10. Group Composition & Size; 11. Growth Data; 12. Habitat Layer; 13. Head-Body Length; 14. Interbirth Interval; 15. Litter size; 16. Litters Per Year; 17. Maximum Longevity; 18. Metabolic Rate; 19. Migratory Behaviour; 20. Mortality Data;  21. Population Density; 22. Ranging Behaviour; 23. Sexual Maturity Age; 24. Teat Number; and 25. Weaning Age. 30 specific variables (see Class IV, Table 1) were extracted from the above data types for PanTHERIA from a total of 94729 data lines (before error checking). Additionally, 4 variables were derived from extracted variables within PanTHERIA and 19 variables were calculated from other spatial data sources (see Class V, Section C).
see http://esapubs.org/archive/ecol/E090/184/metadata.htm for further information.
}
\seealso{
Other rawdata: \code{\link{amphibio}},
  \code{\link{arthropodtraits}}, \code{\link{carabids}},
  \code{\link{heteropteraRaw}}, \code{\link{mammaldiet}},
  \code{\link{passerines}}
}
\keyword{datasets}
