% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{passerines}
\alias{passerines}
\title{Passerine morphology: external measurements of approximately one-quarter of passerine bird speciess}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2366 rows and 27 columns.}
\source{
Ricklefs, R. E. (2017), Passerine morphology: external measurements of approximately one-quarter of passerine bird species. Ecology, 98: 1472. doi: 10.1002/ecy.1783 \url{http://onlinelibrary.wiley.com/doi/10.1002/ecy.1783/suppinfo}
}
\usage{
passerines
}
\seealso{
Other rawdata: \code{\link{amphibio}},
  \code{\link{arthropodtraits}}, \code{\link{carabids}},
  \code{\link{heteropteraRaw}}, \code{\link{mammaldiet}},
  \code{\link{pantheria}}
}
\author{
Robert E. Ricklefs
}
\keyword{datasets}
