% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pulldata}
\alias{pulldata}
\title{Pull trait data from the internet}
\usage{
pulldata(x = NULL)
}
\arguments{
\item{x}{The name of the dataset to source.}
}
\description{
This function sources a recipe for extracting public trait data.
}
\details{
The package 'traitdataform' comes with a collection of recipes for public trait data. These R-scripts
\enumerate{
\item define how to read the file from an online source, i.e. a URL to a txt, xlsx, or a ZIP archive
\item assigns metadata attributes about authorship, license and original publication
\item provide parameters for the standardize() function, i.e. a trait thesaurus, mappings and units.
}

New recipes can be suggested as a pull requrest via the package development page (https://github.com/EcologicalTraitData/traitdataform).
}
\examples{

# to get a list of all available data within the package 
pulldata() 

# to import a dataset
pulldata("carabids")
}
\seealso{
Other rawdata: \code{\link{amniota}},
  \code{\link{amphibio}}, \code{\link{arthropodtraits}},
  \code{\link{carabids}}, \code{\link{heteroptera_raw}},
  \code{\link{mammaldiet}}, \code{\link{pantheria}}
}
\author{
Florian D. Schneider
}
