% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.traitdata.R
\name{rbind.traitdata}
\alias{rbind.traitdata}
\title{Combine trait datasets}
\usage{
\method{rbind}{traitdata}(..., metadata = NULL, datasetID = NULL,
  drop = NULL)
}
\arguments{
\item{...}{two or more objects of class traitdata.}

\item{metadata}{a list of metadata entries which are to be added as
dataset-level information.}

\item{datasetID}{a vector of the same length as number of objects. If \code{NULL}
(default), object names will be returned as ID.}
}
\description{
Function to append objects of class 'traitdata' to each other.
}
\examples{

library(traitdataform)

data(carabids)

dataset1 <- as.traitdata(carabids, 
  taxa = "name_correct", 
  traits = c("body_length", "antenna_length", "metafemur_length"),
  units = "mm",
  keep = c(datasetID = "source_measurement", measurementRemark = "note"), 
  metadata = list(
    bibliographicCitation = attributes(carabids)$citeAs,
    author = "Fons van der Plas", 
    license = "http://creativecommons.org/publicdomain/zero/1.0/"
    )
)

# occurrence table: 

data(heteroptera_raw)

dataset2 <- as.traitdata(heteroptera_raw, 
  taxa = "SpeciesID", 
  traits = c("Body_length", "Body_width", "Body_height", "Thorax_length",
    "Thorax_width", "Head_width", "Eye_width", "Antenna_Seg1", "Antenna_Seg2",
    "Antenna_Seg3", "Antenna_Seg4", "Antenna_Seg5", "Front.Tibia_length",
    "Mid.Tibia_length", "Hind.Tibia_length", "Front.Femur_length",
    "Hind.Femur_length", "Front.Femur_width", "Hind.Femur_width",
    "Rostrum_length", "Rostrum_width", "Wing_length", "Wing_widt"),
  units = "mm", 
  keep = c(sex = "Sex", references = "Source", lifestage = "Wing_development"),
  metadata = list(
    bibliographicCitation = attributes(heteroptera_raw)$citeAs, 
    license = "http://creativecommons.org/publicdomain/zero/1.0/"
    )
)


database <- rbind(dataset1, dataset2)
}
