% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.exploratories.R
\name{standardize.exploratories}
\alias{standardize.exploratories}
\title{Standardize Georeference from Biodiversity Exploratories Plot ID}
\usage{
standardize.exploratories(x, plots = "locationID", user = NULL,
  pswd = NULL, ..., verbose = NULL, warnings = NULL)
}
\arguments{
\item{x}{A traitdata table of class 'traitdata'.}

\item{...}{If input is a rawdata table of type species-trait matrix or occurence table (wide table) then provide parameters according to \code{as.traitdata()}.}
}
\value{
std
}
\description{
Adds columns of georeference to trait-data table if measurements relate to specimens from the Biodiversity Exploratories plots or regions.

This function requires valid credentials for the Biodiversity Exploratories Information System (BExIS)!
}
