% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize trait datasets}
\usage{
standardize(x, ..., verbose = NULL, warnings = NULL)
}
\arguments{
\item{...}{}
}
\value{
std
}
\description{
Standardize trait datasets
}
