% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize.taxonomy}
\alias{standardize.taxonomy}
\title{Standardize scientific names of species.}
\usage{
standardize.taxonomy(x, method = get_gbif_taxonomy, infraspecies = FALSE,
  fuzzy = FALSE, verbose = TRUE, return = c("taxonID",
  "scientificNameStd", "order", "taxonRank"), ...)
}
\arguments{
\item{x}{a traitdata object (as returned by as.traitdata()) or a data table containing at least the column 'scientificName'.}

\item{method}{not functional. Will allow to chose from different sources of taxonomic reference.}

\item{infraspecies}{not functional.}

\item{fuzzy}{disable fuzzy matching if problems with ambiguous species names arise.}

\item{verbose}{has currently no effect.}

\item{return}{a character vector containing the informatoin that should be extracted into the output. Valid entries are the column names returned by function 'get_gbif_taxonomy()'.}
}
\value{
std
}
\description{
Standardize scientific names of species.
}
