% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesaurus.R
\name{thesaurus}
\alias{thesaurus}
\title{Title}
\usage{
thesaurus(x = NULL, replace = NULL, appliestoall = NULL, ...)
}
\arguments{
\item{x}{}

\item{replace}{}

\item{appliestoall}{}

\item{...}{}
}
\value{
std
}
\description{
Title
}
\examples{

traits6 <- thesaurus(
traitID = paste0("A", 1:9),
traitName = c("Body_Size", "Dispersal_ability", "Feeding_guild_short", "Feeding_mode", "Feeding_specialization", "Feeding_tissue", "Feeding_plant_part", "Endophagous_lifestyle", "Stratum_use_short"),
valueType = c("numeric", "numeric", "factor", "factor", "factor", "factor", "factor", "factor", "factor"), 
factorLevels = c("", "", "predator; detritivore; fungivore; herbivore; omnivore", "chewing; extraintestinal; sucking", "monophagous; oligophagous; polyphagous", "", "", "", "" ),
measurementUnit = c("cm","unitless","","","","","","","") 
)

}
