% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{arthropodtraits}
\alias{arthropodtraits}
\title{Functional Arthropod Traits}
\source{
\doi{10.5061/dryad.53ds2};
\href{https://creativecommons.org/publicdomain/zero/1.0/}{Creative Commons
0}. To the extent possible under law, the authors have waived all copyright
and related or neighboring rights to this data.
}
\value{
This is a data object. provides instructions for \code{pulldata()}.
}
\description{
Data from: A summary of eight traits of Coleoptera, Hemiptera, Orthoptera and
Araneae, occurring in grasslands in Germany.
}
\details{
When using this data, please cite the original publication:

\itemize{ \item  Gossner MM, Simons NK, Achtziger R, Blick T, Dorow WHO,
Dziock F, Köhler F, Rabitsch W, Weisser WW (2015) A summary of eight traits
of Coleoptera, Hemiptera, Orthoptera and Araneae, occurring in grasslands
in Germany. Scientific Data 2: 150013.
\doi{10.1038/sdata.2015.13} }

Additionally, please cite the Dryad data package:

\itemize{ \item  Gossner MM, Simons NK, Achtziger R,
Blick T, Dorow WHO, Dziock F, Köhler F, Rabitsch W, Weisser WW (2015) Data
from: A summary of eight traits of Coleoptera, Hemiptera, Orthoptera and
Araneae, occurring in grasslands in Germany. Dryad Digital Repository.
\doi{10.5061/dryad.53ds2} }
}
\seealso{
Other rawdata: 
\code{\link{amniota}},
\code{\link{amphibio}},
\code{\link{carabids}},
\code{\link{heteroptera_raw}},
\code{\link{mammaldiet}},
\code{\link{pantheria}},
\code{\link{pulldata}()}
}
\author{
Gossner MM, Simons NK, Achtziger R, Blick T, Dorow WHO, Dziock F, Köhler F, Rabitsch W, Weisser WW
}
\concept{rawdata}
