% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.traitdata.R
\name{as.traitdata}
\alias{as.traitdata}
\title{Title}
\usage{
as.traitdata(x, traits = NULL, taxa, individuals = NULL,
  occurences = individuals, datasetID = NULL, measurements = NULL,
  units = NULL, keep = NULL, drop = NULL, na.rm = TRUE,
  id.vars = names(x)[names(x) \%in\% keep & !names(x) \%in\% drop],
  mutate = NULL, thesaurus = NULL, ...)
}
\arguments{
\item{x}{}

\item{traits}{}

\item{keep}{}

\item{drop}{}

\item{na.rm}{}

\item{id.vars}{}

\item{...}{}

\item{taxon}{}

\item{occurence}{}

\item{measurement}{}
}
\description{
Title
}
