% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.traittable.R
\name{as.traitdata}
\alias{as.traitdata}
\title{Title}
\usage{
as.traitdata(x, traits, taxa, individuals = NULL, occurences = individual,
  datasetID = NULL, measurements = NULL, units = NULL, keep = NULL,
  drop = NULL, na.rm = TRUE, id.vars = names(x)[names(x) \%in\% keep &
  !names(x) \%in\% drop])
}
\arguments{
\item{x}{}

\item{traits}{}

\item{keep}{}

\item{drop}{}

\item{na.rm}{}

\item{id.vars}{}

\item{taxon}{}

\item{individual}{}

\item{occurence}{}

\item{measurement}{}
}
\description{
Title
}
