% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heteropteraRaw}
\alias{heteropteraRaw}
\title{Morphometric measures of Heteroptera sampled in grasslands across three regions of Germany}
\format{An object of class \code{data.frame} with 425 rows and 34 columns.}
\usage{
heteropteraRaw
}
\keyword{datasets}
