% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammaldiet}
\alias{mammaldiet}
\title{Mammal diet database}
\format{An object of class \code{data.frame} with 5364 rows and 30 columns.}
\usage{
mammaldiet
}
\description{
Establishing macroecological trait datasets: digitalization, extrapolation, and validation of diet preferences in terrestrial mammals worldwide.s
}
\seealso{
Other rawdata: \code{\link{amphibio}},
  \code{\link{arthropodtraits}}, \code{\link{carabids}},
  \code{\link{heteropteraRaw}}, \code{\link{pantheria}},
  \code{\link{passerines}}
}
\keyword{datasets}
