% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.traitdata.R
\name{rbind.traitdata}
\alias{rbind.traitdata}
\title{Title}
\usage{
\method{rbind}{traitdata}(..., metadata = NULL, datasetID = NULL)
}
\arguments{
\item{...}{two or more objects of class traitdata.}

\item{metadata}{a list of metadata entries which are to be added as dataset-level information.}

\item{datasetID}{a vector of the same length as number of objects. If \code{NULL} (default), object names will be returned as ID.}
}
\description{
Title
}
\examples{

library(traitdataform)
data(carabids, arthropodtraits)

dataset1 <- as.traitdata(carabids, 
                        taxa = "name_correct",
                        traits = c("body_length", 
                                   "antenna_length", 
                                   "metafemur_length", 
                                   "eyewidth_corr")
)

dataset3 <- as.traitdata(arthropodtraits,
taxa = "SpeciesID",
traits = c("Body_Size", "Dispersal_ability", "Feeding_guild",
           "Feeding_mode", "Feeding_specialization", 
           "Feeding_tissue", "Feeding_plant_part",
           "Endophagous_lifestyle", "Stratum_use",
           "Stratum_use_short"), 
units = c(Body_Size = "mm", Dispersal_ability = "unitless")
)

newdata <- rbind(dataset1, dataset3, 
               datasetID = c("carabids", "arthropodtraits")
             )
}
