% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize trait datasets}
\usage{
standardize(x, ..., verbose = NULL, warnings = NULL)
}
\arguments{
\item{x}{a traitdata object (as returned by \code{as.traitdata()}) or a data table
containing at least the column `scientificName.}

\item{verbose}{has currently no effect.}
}
\value{
std
}
\description{
wrapper that applies \code{standardize.taxonomy()} and
\code{standardize.traits()} in one go.
}
