% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize trait datasets}
\usage{
standardize(x, ...)
}
\arguments{
\item{x}{a traitdata object (as returned by \code{as.traitdata()}) or a data table
containing at least the column `scientificName.}

\item{...}{parameters as described for \code{standardize.traits()} and \code{standardize.taxonomy()}.}
}
\description{
wrapper that applies \code{standardize.taxonomy()} and
\code{standardize.traits()} in one go.
}
