% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize trait datasets}
\usage{
standardize(x, ...)
}
\arguments{
\item{...}{parameters as described for \code{standardize.traits()} and \code{standardize.taxonomy()}.}
}
\description{
wrapper that applies \code{standardize.taxonomy()} and
\code{standardize.traits()} in one go.
}
\seealso{
Other standardize: \code{\link{standardize_taxa}},
  \code{\link{standardize_traits}}
}
\concept{standardize}
