% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize trait datasets}
\usage{
standardize(x, ...)
}
\arguments{
\item{x}{a traitdata object (as returned by \code{as.traitdata()}) or a data table
containing at least the column `verbatimScientificName.}

\item{...}{parameters as described for \code{standardize.traits()} and
\code{standardize.taxonomy()}.}
}
\value{
A traitdata object with standardized scientific taxon names according
to GBif Backbone taxonomy and standardized trait names according to a
thesaurus, if provided.
}
\description{
wrapper that applies \code{standardize.taxonomy()} and
\code{standardize.traits()} in one go.
}
\seealso{
Other standardize: 
\code{\link{standardize_taxa}()},
\code{\link{standardize_traits}()}
}
\concept{standardize}
