% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize.taxonomy}
\alias{standardize.taxonomy}
\title{Standardize scientific names of species}
\usage{
standardize.taxonomy(x, method = get_gbif_taxonomy, subspecies = TRUE,
  fuzzy = TRUE, verbose = FALSE, return = c("kingdom", "phylum",
  "class", "order", "family"), ...)
}
\arguments{
\item{x}{a traitdata object (as returned by \code{as.traitdata()}) or a data table
containing at least the column `scientificName.}

\item{method}{only option is \code{get_gbif_taxonomy}. In principle, takes any
function that maps the species names in x to produce a taxonomy lookup
table (i.e. mapping user-provided \code{scientificName} to \code{taxonID} and other
taxon-level information). Will allow to chose from different sources of
taxonomic reference.}

\item{subspecies}{logical. If TRUE (default), the given name is resolved to
subspecies epithet, otherwise it will be mapped to species level.}

\item{fuzzy}{if set to \code{FALSE} (default mode), this disables fuzzy matching
if problems with ambiguous species names arise. (see \code{?get_gbif_testing()})}

\item{verbose}{not implemented. If 'FALSE' all messages are suppressed.}

\item{return}{a character vector containing the informatoin that should be
extracted into the output. Valid entries are the column names returned by
function \code{get_gbif_taxonomy()}. See 'Details'.}

\item{...}{parameters to be ignored, forwarded from wrapper function
\code{standardize()}.}
}
\description{
Adds columns to a traitdata object containing accepted species
names and relates to globally unique taxon identifiers via URI.
}
\details{
Taxonomic standardisation is an enormous challenge for biodiversity
data management and research. Constant changes in species and higher taxa,
refinements of phylogenetic trees and changing attribution to original
authors, moving species into other genera or difficulties to place species
into the Linean nomenclature results in highly fluctuent taxonomic
definitions.

As a consequence, there is not one reference for accepted species names and
dependin on the field of resaerch and taxonomic focus other authorities
will be employed.

For reasons of simplicity and because of its high coverage of taxa, the
function \code{standardize.taxonomy()} uses the GBIF Backbone Taxonomy as its
reference system and resolves all provided species names to the accepted
name according to GBIF (resolving misspellings and synonyms in the
process). We invite pull requests to make this function more general and
enable a choice of a taxonomic reference.
}
\examples{


pulldata("carabids")

dataset1 <- as.traitdata(carabids,
  taxa = "name_correct",
  traits = c("body_length", "antenna_length", "metafemur_length"),
  units = "mm",
  keep = c(datasetID = "source_measurement", measurementRemark = "note"),
  metadata = list(
    bibliographicCitation = attributes(carabids)$citeAs,
    author = "Fons van der Plas",
    license = "http://creativecommons.org/publicdomain/zero/1.0/"
    )
)

dataset1Std <- standardize.taxonomy(dataset1)

}
