% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize.taxonomy}
\alias{standardize.taxonomy}
\title{Title}
\usage{
standardize.taxonomy(x, method = get_gbif_taxonomy, infraspecies = FALSE,
  fuzzy = FALSE, verbose = TRUE, return = c("taxonID",
  "scientificNameStd", "kingdom", "taxonRank"), ...)
}
\arguments{
\item{x}{}

\item{infraspecies}{}

\item{fuzzy}{}

\item{verbose}{}

\item{return}{}

\item{reference}{}
}
\value{
std
}
\description{
Title
}
