% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize.traits}
\alias{standardize.traits}
\title{Title}
\usage{
standardize.traits(x, thesaurus, traitmap = NULL, categories = c("No",
  "Yes"), output = "logical")
}
\arguments{
\item{x}{}

\item{thesaurus}{}

\item{map}{}
}
\value{
std
}
\description{
Title
}
