% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize.traits}
\alias{standardize.traits}
\title{Title}
\usage{
standardize.traits(x, map = NULL, thesaurus)
}
\arguments{
\item{x}{}

\item{map}{}

\item{thesaurus}{}
}
\value{
std
}
\description{
Title
}
